+++ hooks/favorite/fetch-favorites.js
@@ -5,10 +5,10 @@ import { fetchFavourites } from 'services/favourites';
 
 const fetcher = (key, token) => fetchFavourites(token);
 
-const useGetUserFavorites = (token, queryConfig = {}) => useQuery(
+const useFetchUserFavorites = (token, queryConfig = {}) => useQuery(
   ['fetch-user-favorites', token],
   fetcher,
   { ...queryConfig },
 );
 
-export default useGetUserFavorites;
+export default useFetchUserFavorites;
