+++ hooks/widget/fetch-widget.js
